lua-yaml: fix yaml encode/decode type inferring
Target behavior is described here:
 http://yaml.org/spec/1.2/spec.html 10.3.2. Tag Resolution.

This patch makes lua-yaml type inferring closer to standard.

The last known difference in type distinguishing related to
the octal numbers. (Because test on number type is performed
by luajit's `tonumber` which do not support octal numbers)
Octal numbers are not supported in other implementations such
as CPython's though.

Related to #3476
Closes #3662 #3583